Weather check using voice with Python.
In this project, I have the idea to create a system with python language where any user can check the weather details of any city using voice. As Alexa from Amazon or Siri from apple, we can use our voice asking question about the current day weather of the city we live in or any other city in the word; then, a voice rebot, which took my name, will responde with detail about the city weather.

Requirements:

First of all, to be able to use my project, we need to have a software that can read python code. I advise to install PyCharm Community Edition which is a free open source. Next,with the python, we have to install some module to allow the the program to use some nedded features. Here are the modules: - speechRecognition module using "pip install SpeechRecognition" in command ligne that allows to take the speech; - pyttsx3 module using command "pip install pyttsx3==2.71" for the return talk; - Pyaudio mudule with "pip install PyAudio" command; if it does not work, go to https://www.lfd.uci.edu/~gohlke/pythonlibs/#pyaudio, and install your python version; - json and requests modules. In addition to these mudules, we also need to get an API key from https://home.openweathermap.org/, but we can use mine given with this source code.